// packages/pf4-component-mapper/src/dual-list-context/dual-list-context.d.ts
export interface SortConfig {
  left: 'asc' | 'desc';
  right: 'asc' | 'desc';
}

export type SetSortConfig = (newSortConfig: SortConfig) => void;

export interface DualListContextValue {
  sortConfig: SortConfig;
  setSortConfig: SetSortConfig;
}

export interface DualListContext {
  value: DualListContextValue;
}

export default DualListContext;